migrate runtime tests to use both kernels
