+++ marimo/_runtime/dataflow.py
@@ -311,6 +311,9 @@ class DirectedGraph:
         for cid in transitive_closure(self, cell_ids):
             self.cells[cid].set_stale(stale=True)
 
+    def get_stale(self) -> set[CellId_t]:
+        return set([cid for cid, cell in self.cells.items() if cell.stale])
+
 
 def transitive_closure(
     graph: DirectedGraph,
