--- tests/_runtime/test_runtime/runtime_data/cell_ui_element.py
@@ -1,17 +1,0 @@
-import marimo
-
-__generated_with = "0.2.8"
-app = marimo.App()
-
-
-@app.cell
-def make_slider():
-    import marimo as mo
-
-    slider = mo.ui.slider(0, 10)
-    slider
-    return (mo, slider)
-
-
-if __name__ == "__main__":
-    app.run()
